renaming things to match library name
