+++ Pipfile
@@ -19,4 +19,4 @@ pytest-cov = "*"
 pytest-sugar = "*"
 
 [packages]
-bcr = {editable = true,path = "."}
+bcr-api = {editable = true,path = "."}
