+++ test/test_credentials.py
@@ -3,7 +3,7 @@ import tempfile
 import unittest
 from pathlib import Path
 
-from bcr.credentials import CredentialsStore
+from bcr_api.credentials import CredentialsStore
 
 ACCESS_TOKEN = "00000000-0000-0000-0000-000000000000"
